:recycle: Moving media manipulation to MediaController
 #4
